jpydzr5-reservation-for-destination/cinema_reservation_system/cinema/views.py
from django.template.response import TemplateResponse
import pendulum


def index(request):
    menu_positions = [
        {"name": "Cennik", "url": "price_list"},
        {"name": "Repertuar", "url": "repertoire"},
        {"name": "Koszyk", "url": "basket"}
    ]
    template = "cinema/index.html"
    return TemplateResponse(request, template, {"menu_positions": menu_positions})


# Create your views here.
def basket(request):
    template = "cinema/basket.html"
    message = {"message": "OK!"}
    return TemplateResponse(request, template, message)


def repertoire(request):
    template = "cinema/repertoire.html"
    start_day = pendulum.now("Europe/Warsaw")
    seven_days_forward = {}
    for day in range(1, 8):
        start_day = start_day.add(days=1)
        seven_days_forward[start_day.format("YYYY-MM-DD")] = start_day.format("dddd", locale="pl")
    context = {"days": seven_days_forward}
    day = request.GET.get("date")
    return TemplateResponse(request, template, context)


# Create your views here.
def price_list(request):
    template = "cinema/price_list.html"
    message = {"message": "OK!"}
    return TemplateResponse(request, template, message)